# openapi.yaml
openapi: 3.0.3
info:
  title: ExecFlex API
  description: |
    ExecFlex API - Combined API Server for executive matching and voice interactions.
    
    This API handles both web-based operations (role posting, matching, introductions) 
    and voice/telephony features powered by Ai-dan, an AI voice agent.
    
    ## Features
    - Executive candidate matching
    - Role posting and management
    - Email introductions
    - Voice calls via Twilio (Ai-dan)
    
    ## Authentication
    Most endpoints require JWT authentication via Supabase Auth. Include the JWT token in the Authorization header:
    `Authorization: Bearer <supabase_jwt_token>`
    
    Twilio webhook endpoints (`/voice/*`) are public and called by Twilio (signature verification recommended for production).
  version: 1.0.0
  contact:
    name: ExecFlex API Support
    url: https://execflex.com
servers:
  - url: https://execflex-backend-1.onrender.com
    description: Production server (Render)
  - url: http://localhost:5001
    description: Local development server
...
paths:
...
  /onboarding/enqueue:
    post:
      tags:
        - Onboarding
      summary: Manually trigger onboarding for a user (admin-only)
      description: |
        Manually trigger onboarding for a user. This endpoint requires authentication AND admin role.
        Used for manual/admin triggers, testing, or re-triggering onboarding.
        
        **Note:** Automatic onboarding is handled by database trigger on signup,
        so this endpoint is primarily for admin operations or edge cases.
        
        **Requires:** JWT authentication + admin role in role_assignments table.
      operationId: enqueueOnboarding
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
              properties:
                user_id:
                  type: string
                  format: uuid
                  description: The user_id to initialize onboarding for
                  example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Onboarding triggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  job_id:
                    type: string
                    format: uuid
                    description: ID of the created outbound call job
                  thread_id:
                    type: string
                    format: uuid
                    description: ID of the created thread
                  interaction_id:
                    type: string
                    format: uuid
                    nullable: true
                    description: ID of the created interaction
                  status:
                    type: string
                    enum: [queued, running, succeeded, failed]
                    description: Status of the onboarding job
              example:
                ok: true
                job_id: "123e4567-e89b-12d3-a456-426614174000"
                thread_id: "456e7890-e89b-12d3-a456-426614174001"
                interaction_id: "789e0123-e89b-12d3-a456-426614174002"
                status: "queued"
        '400':
          description: Missing user_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /onboarding/process-jobs:
    post:
      tags:
        - Onboarding
      summary: Manually trigger job processing (optional)
      description: |
        Endpoint to trigger job processing for HTTP-based cron/scheduled tasks.
        
        **Note:** This endpoint is optional. If using Render Background Workers,
        you can run `python -m workers.call_dispatcher` directly instead.
        
        Processes queued outbound call jobs and initiates Twilio calls.
      operationId: processJobs
      parameters:
        - name: limit
          in: query
          description: Maximum number of jobs to process
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Jobs processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  processed:
                    type: integer
                    description: Number of jobs processed
                  message:
                    type: string
              example:
                ok: true
                processed: 3
                message: "Processed 3 jobs"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /onboarding/status:
    post:
      tags:
        - Onboarding
      summary: DEPRECATED - Twilio status callback (use /voice/status)
      description: |
        **DEPRECATED:** This endpoint is kept for backward compatibility only.
        Use `/voice/status` instead.
        
        Twilio status callback webhook for onboarding calls.
        Updates job and interaction records with call status.
      operationId: onboardingStatus
      deprecated: true
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                CallSid:
                  type: string
                  description: Twilio Call SID
                  required: true
                CallStatus:
                  type: string
                  description: Call status
                  enum: [queued, ringing, in-progress, completed, failed, busy, no-answer, canceled]
                CallDuration:
                  type: string
                  description: Call duration in seconds
                From:
                  type: string
                  description: Caller's phone number
                To:
                  type: string
                  description: Called phone number
      responses:
        '200':
          description: Status update processed
          content:
            text/plain:
              schema:
                type: string
                example: "OK"
        '400':
          description: Missing CallSid
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase JWT token from authenticated user session.
        Get the token from `supabase.auth.getSession()` in the frontend.
        Include in header: `Authorization: Bearer <token>`
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - ok
        - error
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
          example: "Missing required fields: email, name"
...